A basic router for the Aurelia app.
