+++ aurelia/src/app.js
@@ -1,20 +1,13 @@
-import { inject } from 'aurelia-framework';
-
-import { DataService } from './services';
-
-@inject(DataService)
 export class App {
-  constructor(dataService) {
-    this.dataService = dataService;
-  }
+  constructor() {}
 
-  filterComedian(filterText) {
-    let lowerFilterText = filterText.toLowerCase();
-    this.comedians = this.dataService.getComedians()
-      .filter(c => c.fullName.toLowerCase().includes(lowerFilterText));
-  }
-
-  activate() {
-    this.comedians = this.dataService.getComedians();
+  configureRouter(config, router) {
+    this.router = router;
+    config.title = 'Aurelia';
+    config.map([
+      { route: ['', 'home'],       name: 'home',       moduleId: 'comedians' },
+      { route: 'comedians',            name: 'comedians',      moduleId: 'comedians',   nav: true },
+      { route: 'comedians/:id/detail', name: 'comedianDetail', moduleId: 'comedianDetail' }
+    ]);
   }
 }
