+++ aurelia/src/comedianDetail.html
@@ -1,0 +1,5 @@
+<template>
+    <h1> ${comedian.fullName}</h1>
+    <h2> ${ comedian.group.name } </h2>
+    <h3><a target='_new' href.bind='comedian.url'>${ comedian.url }</h3>
+</template>
