+++ aurelia/src/comedianDetail.js
@@ -1,0 +1,14 @@
+import { inject } from 'aurelia-framework';
+
+import { DataService } from './services';
+
+@inject(DataService)
+export class Comedians {
+  constructor(dataService) {
+    this.dataService = dataService;
+  }
+
+  activate(params) {
+    this.comedian = this.dataService.getComedian(params.id);
+  }
+}
